# drivers/vnfm/svnfm/ericsson/vnfmdriver/VnfmdriverService/service/src/main/resources/swagger/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Ericsson Vnfm Driver API",
    "version" : "1.0.0",
    "title" : "Ericsson Vnfm Driver API"
  },
  "host" : "ericsson-vnfm",
  "tags" : [ {
    "name" : "v1"
  } ],
  "schemes" : [ "https", "http" ],
  "paths" : {
    "/openoapi/ericsson-vnfm/v1/{vnfmId}/jobs/{jobid}&responseId={responseId}" : {
      "get" : {
        "tags" : [ "v1" ],
        "summary" : "Get operation status",
        "description" : "",
        "operationId" : "getOperationStatusJson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "resp",
          "in" : "query",
          "description" : "response",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "jobid",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "responseId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "language-option",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VNFServiceRestAPI"
              }
            }
          },
          "404" : {
            "description" : "VnfmDriver Resource Not Found."
          },
          "500" : {
            "description" : "VnfmDriver resource failed to process the request."
          }
        }
      }
    },
    "/openoapi/ericsson-vnfm/v1/{vnfmId}/vnfs" : {
      "post" : {
        "tags" : [ "v1" ],
        "summary" : "Instantiate VNF",
        "description" : "",
        "operationId" : "instantiateVNFJson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "req",
          "in" : "query",
          "description" : "request",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resp",
          "in" : "query",
          "description" : "response",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "language-option",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VNFServiceRestAPI"
              }
            }
          },
          "404" : {
            "description" : "VnfmDriver Resource Not Found."
          },
          "500" : {
            "description" : "VnfmDriver resource failed to process the request."
          }
        }
      }
    },
    "/openoapi/ericsson-vnfm/v1/{vnfmId}/vnfs/{vnfInstanceId}" : {
      "get" : {
        "tags" : [ "v1" ],
        "summary" : "Query VNF",
        "description" : "",
        "operationId" : "queryVNFJson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "resp",
          "in" : "query",
          "description" : "response",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vnfInstanceId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "language-option",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VNFServiceRestAPI"
              }
            }
          },
          "404" : {
            "description" : "VnfmDriver Resource Not Found."
          },
          "500" : {
            "description" : "VnfmDriver resource failed to process the request."
          }
        }
      }
    },
    "/openoapi/ericsson-vnfm/v1/{vnfmId}/vnfs/{vnfInstanceId}/terminate" : {
      "post" : {
        "tags" : [ "v1" ],
        "summary" : "Terminate VNF",
        "description" : "",
        "operationId" : "terminateVNFJson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "req",
          "in" : "query",
          "description" : "request",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resp",
          "in" : "query",
          "description" : "response",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vnfInstanceId",
          "in" : "path",
          "description" : "URLParam",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "language-option",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VNFServiceRestAPI"
              }
            }
          },
          "404" : {
            "description" : "VnfmDriver Resource Not Found."
          },
          "500" : {
            "description" : "VnfmDriver resource failed to process the request."
          }
        }
      }
    },
    "/openoapi/nslcm/v1/grantvnf" : {
      "post" : {
        "tags" : [ "v1" ],
        "summary" : "Grant VNF lifecycle operation",
        "description" : "",
        "operationId" : "grantVNFJson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "req",
          "in" : "query",
          "description" : "request",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resp",
          "in" : "query",
          "description" : "response",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "language-option",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/NSLCMServiceRestAPI"
              }
            }
          },
          "404" : {
            "description" : "VnfmDriver Resource Not Found."
          },
          "500" : {
            "description" : "VnfmDriver resource failed to process the request."
          }
        }
      }
    }
  },
  "definitions" : {
    "NSLCMServiceRestAPI" : {
      "type" : "object"
    },
    "VNFServiceRestAPI" : {
      "type" : "object",
      "properties" : {
        "apiDoc" : {
          "type" : "string"
        }
      }
    }
  }
}